Доработанное ДЗ 13

+++ src/ru/innopolis/homework013/Utils/InputData.java
@@ -4,15 +4,13 @@ public class InputData {
 
     public InputData() {
     }
-
     public static Integer parseCount(String inputNewData) throws NumberFormatException {
         try {
-            return (Integer) Integer.parseInt(inputNewData);
+            return  Integer.parseInt(inputNewData);
         } catch (NumberFormatException error) {
             throw new NumberFormatException("Невалидное значение" + error);
         }
     }
-
     public static Integer validateCount(String inputNewData) throws NumberFormatException {
         try {
             return parseCount(inputNewData);
@@ -20,15 +18,13 @@ public class InputData {
             throw new NumberFormatException("Невалидное значение" + error);
         }
     }
-
     public static Double parseNumber(String inputNewData) throws NumberFormatException {
         try {
-            return (Double) Double.parseDouble(inputNewData);
+            return  Double.parseDouble(inputNewData);
         } catch (NumberFormatException error) {
             throw new NumberFormatException("Невалидное значение" + error);
         }
     }
-
     public static Double validateNumber(String inputNewData) throws NumberFormatException {
         try {
             return parseNumber(inputNewData);
